Redesign German to show "Blue Hand"
We have the go ahead to move forwards with the German site! It has been
a while but the German authorities have given us the go-ahead with the
design of the login/enrol screen.

Next steps:

Configure the login/register page as per the attached version – I have
attached the PDF and the packaged InDesign files.
Update footer as per the attached, but change the month to May in German
(Mai):
VIZAMYL DE Version 1.1 – Mai 2020

©2020 General Electric Company – Alle Rechte vorbehlten.

Replace these three videos:
Introduction video

M3_P3

M3_P27

The new ones can be downloaded here: https://we.tl/t-5Vfx6soNOy

If you’re not sure on the which pages these videos are from I can find
the url. But hopefully the code on the file name is enough.

Remove the blue hand from footer and last page. It is only needed on the
register/register page

We need to keep it in the Staging site for now so they can review and
approve it on the staging site before it goes live.

Please can you let me know when you think you’ll be able to get this
done? There isn’t any particular pressure from the client but I will
give them an idea of turn around time.

Would by the end of next week be feasible?

If you have any questions, just let me know.

Thank you,

Lily