Add a Belian gatekeeper page
Hi Andrew,

The client has come back to me with to confirm the updates to the
Belgium site.

Please add Belgium to the homepage. I suggest arranging the countries in
3 columns now, as 4, 4, 3 as we now have 11 countries Please create the
  gatekeeper page for Belgium, with your suggestion of having the
buttons which would like to the English or French access page The client
has decided not to proceed with adding the footer bar to the new page. I
think they realised that it would be complicated and unnecessary so
didn’t want to do it.

Please can you let me know if you have any questions around this.

We have budget to cover your time for this so you can crack on when you
get an opportunity.

Many thanks,
Lily