Change address and phone number
Lily Yates
14:09 (2 hours ago)
to me

The client has reviewed the landing page for Belgium and has two small
text changes to make to it before it can go live.

In the paragraph under ‘Reporting adverse events:’ there is a an address
which needs updating.

FROM: Eurostation II – Place Victor Horta 40/40 – 1060 Brussels

TO: PO Box 97 – 1000 Brussels, Madou

In the paragraph under that for Luxembourg can we change this phone
number

FROM: (+35)2 247 85593

TO: (+352) 2478 5592

Please can you make this update onto the Staging site. I will get the
client approval before it goes live.