Disable video for English and German
Hi Andrew,

I have just heard from the client that the changes we have made to the
Austrian part of the readvizamyl site have been approved and can go
live.

This is for Austria only – not Germany or the Belgium/Luxembourg site.

Please can you make the Austrian site live and let me know when it is
ready so I can inform the client.

Thank you,

Lily